+++ Changelog.md
@@ -3,5 +3,6 @@
 ## 0.0.1 - release
 
 ### Feature
+ - improve time matching logic in parser options
  - add file driver for parser
  - add parser